+++ extensions/google-meet/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Google Meet Changelog
 
+## [New Command] - {PR_MERGE_DATE}
+- 
+
 ## [Improvement] - 2024-10-22
 
 - Add delay before reading meeting URL from browser